Add interim claim info relationship to claims
Requirement is for the users to be able to, for a final claim,
(re)submit the dates for a interim warrant fee (or to confirm the
execution date).

NOTE: After some offline discussion became clear that the real intent is to
be able to relate final claims with interim claims. Due to lack of
requirements on how they will be related we are going with this approach
instead.