Zero cost redeterminations (#2384)
* Add feature test for non-display of empty refusal assessment

Also, extend feature definitions by allowing creation of a
redetermination claim and an explicit check of message text contents

* Prevent setting assessment to zero on post-assessment refusal

The `set_amount_assessed_zero!` was called on all refuse and reject
transitions.  This patch prevents it being called if the claim has
previously been assessed.

* Clarify unit test

* Tidy of tests to better use the event

Now it tests the execution of the event, rather then just the outcome

* Update feature test

After rebasing in master the claim factory had been updated and
the form_step needs to be manually set to ensure the validation
runs as it used to.  This will need to be updated in the future,
hence the TODO comment, replicating the use in the test above.