Define endpoint to return a list of establishments
NOTE 1: The initial solution was to not have this endpoint at all but
instead try to make use of any existent register for the data we require
to populate the travel expenses location field.

The current registers either don't exist for a particular type of
establishment or do not have the necessary data (e.g. postcode) we
require.

NOTE 2: The implemented endpoint returns a JSON list of establishments that
includes the name, category and postcode. The list can be filtered by
supplying a given category.