Do not display amount in CYC page for fees that don't require it
CYC (check your claim) page

NOTE: fees like PPE under scheme 10 do not require an amount (whereas
before they did). so for those cases, the same way that they are not
require to input an amount in the form for that fee, the CYC page does
not display the amount for it.