Handle case uplift fixed fees
By calculating unit prices for
case uplifts we remove the need to
change the underlying data structures
and modify the sidebars functionality,
both of which will involve significant work
on back and front ends respectively.

This is not an ideal solution as it leaves
the data structures in a state which is
fundamentally different from that required
for a simple calculation approach.