Add a service to generate and persist the requested reports
NOTE 1: This currently uses the same behaviour as the existent MI report
which is to store the report in the database directly. Ideally we will
replace this behaviour with a more sustainable approach where the report
files are stored in S3 storage (or similar) and the users will have
download the report from there.

NOTE 2: I've choosen to separate some elements of the
generation/persistance/notification as the original MI report service
was doing too much. This new approach has:
- a service object that can perform the appropriate query to retrieve
the data from the database
- a service object that allows us to export the retrieved data into a
format of our choice (for now CSV only), without persisting any data,
just returning those results
- a service object that has all the business logic of creating a report
request from the user, setting its start and completion date and the
content of the actual report