Fix advocate categories on fixed fee page
Found that for a new AGFS claim the advocate categegories
on the fixed fee pages were showing the scheme 9 categories.

Not sure if this came out of the recent refactor of the fee reform
code but the fetch eligible advocate categories services was
still relying solely on the `fee_scheme:string` of `'fee_reforms'`.