Updates to fix the speed / timing issues on the page
Each expense has a number of form elements, the visibility of each element is controlled by the JS each time the expense type is changed.

This show / hide process was very slow, causing to side bar not to ‘see’ some values when the page is loaded.

In this commit:
- remove the need for blocks to reload when page is loaded
- state updates are now done after the element is detached from the DOM
- cost per mile refactored to account for car / bike state

==============================================
Add table to store establishments

To store basic information such as the establishment name, its category
(e.g. prison, hospital) and their postcode.

Add service to seed prison establishments

Add seeds for hospital establishments

Add seed data for magistrates court establishments

Add seeds for crown court establishments

Add log scope to each establishment category seeder

======================================================

Define endpoint to return a list of establishments

NOTE 1: The initial solution was to not have this endpoint at all but
instead try to make use of any existent register for the data we require
to populate the travel expenses location field.

The current registers either don't exist for a particular type of
establishment or do not have the necessary data (e.g. postcode) we
require.

NOTE 2: The implemented endpoint returns a JSON list of establishments that
includes the name, category and postcode. The list can be filtered by
supplying a given category.

======================================================

Remove Expense versioning

There were currently two versions being managed by the application even
though production data suggest only V2 is in use.

This removes the version management as its no longer necessary and just
adds extra unnecessary entropy to the application's business logic.

======================================================

Add location type to expenses

Given the current model for expenses tracks locations as a string, we
will need some extra information stored in order to perform calculations
in the future for the distance between the provider location and the
destination.