Default eligible advocate categories to all for nil fee scheme
A fixed fee case type claim will not be able to
determine fee scheme from offence at point of claim
creation because it does not require an offence. When
created via the API it does not have a rep
order either so fee scheme is indeterminable
using current logic and models.