js-block init process
The Sidebar will maintain an array of `blocks` that is used to for calculation. This array was created from scrach each time a fee / expense / item  was added to the page.

The recreation of blocks is causeing issues with events and could overload an already congested page.

In this commit:
- Introduce `.fx-do-init` class that will be removed after init is compelted
- Cocoon will duplicate with that class intact and init will of the new block kicks off
- Applies to all AGFS + LGFS `.js-block`