Laa fee calculator client integration (#2427)
* Add laa-fee-calculator-client 0.1.1

Use published gem instead of github source code.

* Add ajax and controller action

* Add fee calculator service

* Refactor fee calculator service

* Convert to using JSON and ajax directly

Simpler/clear to rely on javascript in the ajax call itself
rather than use ajax and rails preprocessed partials.

* Rename selectors and change calculated message result

* Make advocate type passable options

* Refactor calculator controller action and service

Set required objects inside the service, and use
options rather than individual args.

* Add unit price calls

Add a unit price controller action
service and ajax call.

* Remove JS format response for fee calculation actions

Was used initially before relying on
ajax alone.

* Add service object response

More idiomatic for services to return
a kind of response object, and to include
a data and errors attribute.

* Handle response errors

Add a catchall for errors raised by the API,
client, or CCCD to ensure user is aware that
unit price could not be determined.

* Apply gov uk error styling

* Handle case uplift fixed fees

By calculating unit prices for
case uplifts we remove the need to
change the underlying data structures
and modify the sidebars functionality,
both of which will involve significant work
on back and front ends respectively.

This is not an ideal solution as it leaves
the data structures in a state which is
fundamentally different from that required
for a simple calculation approach.

* Refactor to use jquery deferred object

Prepare for enabling recalculation on
all or one fixed fee by separating the
ajax call to its own function and separating
the events to one requiring a change to all
fee unit prices (advocate type change) and one for
recalculating just a specific fees unit price
(fee type change).

* Handle multiple fixed fees.

* Configure fee calculator for deployed environments

Use client gems internal default of the
API dev enviroment for now.

* Refactor unit price service

Promote case uplift handling to superclass
and simplify call methods.

* Handle case uplift base fee changes

When the number of days is changed for a "primary"
fixed fee any case uplifts related to that fixed fee
must recalculate their value taking into account the
change to its "parent".

* Handle fixed fee removal impact on fee calc

Some fixed fee records are inter-dependant
and therefore removing one may inpact the other.

* Isolate ajax data hash construction

* Amend case upliftable

add missing mapping and wrap
constants with methods for use
in other contexts.

* Refactor calculate service

* Rename net amount to rate

* Move case numbers field below fee type

* Add "Number of defendant uplift" unit price calculation

* Add #calculate_unit_price controller spec

* Add fixed fee net amount read-only field

To display the rate * quantity value per fee
in line with designs [CBO-294](https://dsdmoj.atlassian.net/browse/CBO-294)

* Add fee calculator service specs

* Bump laa-fee-calculator-client version and rerecord VCR cassettes

* Ignore brakeman warning

This has been broken for a while and prevents
a full test suite run locally. Ignoring for now.

* Modify factories to allow fixed rep order dates

Current factories result in two rep orders
per defendant and of 400 and 380 days ago.

This modifies the factory helpers and
transient attributes to allow creation
of one defendant with a scheme 9 rep
order using.

```
  let(:claim) do
    create(:draft_claim,
      create_defendant_and_rep_order: false,
      create_defendant_and_rep_order_for_scheme_9: true,
      case_type: case_type, offence: offence
    )
  end
```

TODO: needs wrapping up more neatly and a separate
factory for fee calculator claims generally.

* Prevent inappropriate errors and make rate read-only

Following designs the rate, is available, should not
be amendable by the user.

Also, do not display an calc error for no fee type
being selected, as is default for a new page and
new fixed fee.

* Modify existing fixed fee features in line with fee calculation

These features now need to record VCR cassettes for the fee
calculator API responses expected and check expected unit prices
are returned. Also modified one from being contempt which does not
logically have number of cases (or defendant) uplift fee (according
to the calculator itself).

* Update claim_form_page for in-browser testing

* Add a defendant uplift step definition

* Add a stand alone fee_calculator feature for defendant uplift

* Call the defendant_uplift_unit_price method explicitly

Commented to explain why

* Add a TODO to remove console.log debug

* Add step definitions for handling calculator error states

* Stub out calculator calls in features

This can be extended to handle other error types

* Add feature for when calculator offline

No VCR as webmock stubs aout a failing request

* remove @fee_calc_vcr tag

This test uses the webmock failure tag instead and should not be
included in VCR runs

* Update error message in line with CBO-306

* Replace error display with form-hint

* Clear the rate when the calculator cannot return one

* Add `rate_calculated` to fees

This allows us to handle state when a fee is calculated
We can make fields read only on subsequent page loads and, in
the future, display markers for case workers that fess have been
calculated internally

* Handle the new rate_calculated field

* Sort Advocate claims fixed_fees by the order created

* Refactor the adp_text_field

Reduce the complexity score by removing duplication

* Remove the caseNumbersInput.focus

TODO : Speak to @jsugarman about why the focus was added as it removes
the error display by selecting the field

* Add test for adp_text_field.readonly generator

* Update fixed fee view to call refactored adp_text_field

* Remove superfluous gem

* remove required ostruct

* Move rate_calculated view logic into fee object