Chrome fixed fee calculation feature (#2470)
Add window focus to sign in step for chrome driver

By default the Selenium chrome browser window
will be in the background and hard to find which can
lead to confusion. This will bring it to the foreground.

In addition, using the send keys triggers a change event
which seems to be required by chrome browser.