Add safe navigator to collection of case numbers
Currently it is possible to, as a separate bug,
to step through a path of form submission
which ends up with a valid claim submission but
with a fixed fee case uplift that has no value
and no additional case numbers.

Until this bug is resolved the addition of
the safe navigator will prevent a 500 on
injection of such a claim.