Handle elected case not proceeded fixed fee calculation (#2486)
Add elected case not proceeded fixed fee calculation

Although ENP is treated as a fixed fee in CCCD it
actually maps to an advocate fee with a fixed price
in the calculator (AGFS_FEE).

The prices endpoint JSON for the price of this
fee stores the value in a `fixed_fee` attribute
instead of the usual `fee_per_unit` attribute.

The amount is not affected by either advocate
category or unit multipliers. This last will need
handling by JS in the current CCCD configuration.