Use claim's apply_vat flag for expenses
A bug was raised by a vender in which they
were submitting final advocate claims via
the API explicitly setting the apply_vat=true,
but there user was not vat_registered. This
showed up a bug whereby fees where respecting
the claims apply_vat flag but expenses were
not.

In addition the Sidebar calculations were respecting
the claims apply vat flag but the backend was
not. This meant the sidebar would update expense VAT when a
fee or expense was edited but moving to the next
page would display the persisted vat in sidebar - which
for expenses had not been saved.