Have ajax pass destination postcode
Previously this was returning the locations
`name` which leads to google api picking a central
spot in the given location.

This changes the front end to return the pre-existing
postcode data-attribute (itself created from the
establishment model's postcode attribute). In addition
the service object will validate the destination param
as being a valid postcode.

NOTE: Solihull Magistrates Court's postcode is UNKNOWN
at the moment and so distance is incalculable.