CBO 296 AGFS miscellaneous fees (#2491)
* Create test for calculating AGFS Misc fees
* Update the Misc Fee views
Remove line space from parent
Add required fields for misc fees to be calculated
* Update fee calculation JS
Rebase some of the methods to reduce duplication of fixed and misc fee
variations
* Add values for misc fee calculation hint text
* Update the fee calculator Price method
It now returns a `unit` that can be passed back to the view, this is
used to update the hint text for quantity
* Update the fee calculator UnitPrice
It now returns the price passed back to the calculate parent, this is
used to update the hint text for quantity
* Update the fee calculator Calculate class
It now returns the amount _and unit_ if a price is returned from the
amount method in unit_price, this is used to update the hint text
for quantity
* Update DefendantUpliftable
This allows the handling of uplifting defendant based misc fees
* Update the feature/advocate_claim_steps
This allows re-use of steps with variable misc/fixed fee switches
* Re-factor the existing features to use the new steps
This also allows calculation of misc fees, rather than manually entered
fields
* Re-run the VCR cassettes
Ensure the VCR Cassettes are up-to-date, this could be adding or
re-running previous ones that now need misc fee API submissions too
* Remove is_rate_calculated? method
Replace references with rate_calculated? as generated on boolean fields
automatically by Rails
* Remove fee-type specific tags from the views
e.g. replace `js-misc-fee-calculator-fee-type` with `js-fee-calculator-fee-type`
* refactor FeeCalculator Javascript
* Combine/remove methods to allow fixed/misc fees to call the same code
* Remove duplication of DOM traversal for finding elements in populateNetAmount
* Update the application JS to call the new methods
Also removes the, now superfluous, duplication of cocoon::after-insert of fee code
* refactor JS for feeQuantityChange event
* Remove specific classes for fixed and misc fees oon the views
* set application.js to duplicate the event on the genric class
* Add hide_hint function to adp_text_field
This will allow us to hide a hint by default, until we set it with JS
and then show it in Fee Calculation
* Call the new hide_hint version of adp_text_field on misc_fees view
* Update the base fee to order by ID
This should allow _all_fees to be displayed in the order they were added
* Update the JS to re-call the API on page load
This is designed to update:
* the hints to display the correct units
* re-calculate fees in the event of changes to the previous pages, e.g.
advocate category, offence, etc
* Update hint text for defendants
* Extend test coverage of the unit_price endpoint
Add misc_fee and misc_fee uplift tests