Refine the SoftlyDeletable handling
Only update the deleted_at attribute of claims on soft_delete

This was required as Providers were unable to delete claims that were in
an invalid state, `update` forces re-validation, `update_attribute` does
not.

I was reluctant to turn this on for _all_ models as the scope of the
change was for claims, and errors on other models had not occurred. Yet