Add API integration test for LGFS fixed fees
To duplicate current API vendor submissions
this test POSTs a litigator final claim fixed
fee using the old/current `amount` attribute
having a value.

This will result in a valid draft fixed fee
but the amount will be ignored/nil'd as a result
of the new calcuated fixed fee approach.