Amend CCCD API to adapt LGFS fixed fees
Since we are changing what values are expected
from the API for LGFS fixed fees it has been
decided to adapt POSTs from consumers so that
they fit the new quantity and rate requirement.

This is done simply by defaulting quantity to 1 and
the rate to the amount (posted by the consumer).