Refactor fee calc javascript to use namespacing/sub-modules
To reflect the class-like structure/nature
of fee calculation. There are now two fundamental
fee calculation call methods - one to calcululate
unit prices via the price endpoint of the api
and the other to calculate an entire fee
via the calculate endpoint of the API. The
backends structure reflects this and it
woud be good to have a similar structure
on the frontend.