Add VCR config to handle chrome browser run
When VCR cassettes are used in a feature test
and the test is run using chrome driver, request
to 127.0.0.1:9515 are being made. VCR tries
to and fails to record these, as it should.

This VCR config change explicitly ignores these
requests and enables visible chrome browser runs
to function.

e.g.
```
$ BROWSER=chrome features/fee_calculator/litigator/fixed_fee_calculator.feature
```