Add missing LGFS API integration tests
Adds the missing graduated fee, interim fee
and transfer fee claim API integration tests.

These are currently only tested via the smoke test
but are valuable as a more immediate form of feedback/test
to ensure we do not break the API and to assist in debugging
bug/issue reports from API consumers.