Use unpresented fee amount for field value
This field was using a presented fee amount that
resulted in, for example, £20.00 in the
value attribute of the tag. Does not
have an impact on value saved when it is
changed but if fee calc is deactivated
(if we need too, by commeting .calculated-grad-fee
css class out) it will not render the
presented amount as the value.