Use unpresented fee rate for LGFS fixed fee rate
This field was using a presented fee rate that
resulted in, for example, £20.00 in the
value attribute of the tag. Does not
have an impact on value saved as it is
recalculated but would have impact
(if we need too, by commeting .calculated-grad-fee
css class out) it will not render the
presented rate as the value.