Add factory helper for scheme dates
Scheme dates have important validation implications
and the dates for commencment of a given scheme should
be shared centrally. RSpec, Cucumber, FactoryBot and Smoke
tests all need access to these but their config/setup is different.

for FactoryBot: load factory helpers as part of cucumber config
in the env.rb file, so we do not have to explcitly
include them where needed. Also allows a single
location for both rspec and cucumber factorybot
helpers to be stored.