Fix amoeba/cloner
- remove monkey patch infinite loop and rely on amoebas propate method
- "Redraft" error translation tweak
- Amend cloner spec inline with advocate interim claims
- fix logging text spec
- Add amoeba config tests

The loop in the duplicable module was causing an infinite
loop during deployment/startup of app. The superclass for
fee was returning Fee::BaseFee infinitely. Still have not
determined why this happens on outs envs but not locally.

Instead rely on amoebas own propagate method that is intended
for STI models.