Fix count of claims in claim list
Seems to have been cause by rails 5.1 upgrade
where the alias used in the order by clause is
stripped when applying a count/size activerecord
operation on the statement.

Still looking for reason it has failed now but
by unscoping the order (which is irrelevant) to
a count operation anyway, it fixes it.