Make basic fee mappings available to fee calculator services
In order to distinguish between different basic fee
types while still mapping to the same API fee types
I have extended the basic fee adapter to behave
much like the fixed and misc fee adapters.

This requires the basic fee adapter to be instantiable
both with a claim object and without one, since for fee
calculation we just want to reuse existing functionality
of the fee calculation services related to fee type mappings.
At the same time we want to preserve the functionality of
this adapter for DI purposes.