Set a default api key in Rails.secrets
This was a valid key when the VCR cassette was generated but has now
been expired.  If new tests are written, this should be replaced as
publishing this in the open is probably a bad idea.

New, free, keys can be generated at https://currencylayer.com
and yours should be stored in `.env.development`