Create currency conversion service
This will allow us to generate quarterly volumes

- Set up VCR to prevent calls being made to the live
  server that will evenntually charge!

Set a default api key in Rails.secrets
This was a valid key when the VCR cassette was recorded but has now
been expired.  If new tests are written, this should be replaced as
publishing this in the open is probably a bad idea.

New, free, keys can be generated at https://currencylayer.com
and yours should be stored in `.env.development`