Add API endpoints for the QuarterlyVolume report
Long term, ideally, this would be authomated as a scheduled task too

In the meantime, there is a GET endpoint that generates the required
figures from the supplied data:
- the start date of the quarter
- the three monthly USD dollar hosting values for the quarter

It stores the output in a tmp file and dipslays he value for the user
to confirm the data

Then, there is a POST endpoint that reads the tmp file and sends it
to the performace platform.  On successful post, it deletes the tmp
file.