Add safe operator to translation scope construction
LGFS Claims with a rep order date before scheme 8s
offical start date do not have a fee scheme. This
causes the view rendering to fail when attempting
to contruct a translation for the page title.

The logic (no fee scheme for claim) may otherwise
be correct since fee calc and other things for LGFS
cannot reliably be applied for such claims??