CBO-494: Performance Platform Integration (#2626)
* Implement the PerformancePlatform Class

This is deliberately set-up so that it can be gemified

* Configuration for the performance platform class

* Create the PerformancePlatform.TransactionsByChannel report

Include a scheduled task to call it

* Create currency conversion service

This will allow us to generate quarterly volumes

- Set up VCR to prevent calls being made to the live
  server that will evenntually charge!

Set a default api key in Rails.secrets
This was a valid key when the VCR cassette was recorded but has now
been expired.  If new tests are written, this should be replaced as
publishing this in the open is probably a bad idea.

New, free, keys can be generated at https://currencylayer.com
and yours should be stored in `.env.development`

* Add a Claims::Count service method

This allows a simple, re-useable, reliable class to get claim counts
for generating performance platform stats

* Create the PerformancePlatform.QuarterlyVolume report

* Add a generator for the QuarterlyVolume report

* Add API endpoints for the QuarterlyVolume report

Long term, ideally, this would be authomated as a scheduled task too

In the meantime, there is a GET endpoint that generates the required
figures from the supplied data:
- the start date of the quarter
- the three monthly USD dollar hosting values for the quarter

It stores the output in a tmp file and dipslays he value for the user
to confirm the data

Then, there is a POST endpoint that reads the tmp file and sends it
to the performace platform.  On successful post, it deletes the tmp
file.