Add advocate supplementary claim type and web flow
Spike to demo the idea of an entirely
new claim type to handle miscellaneous
fees not claimed as part of a final claim.

Users previously worked around the need
to do this by claiming the kind of fees
as part of a final claim but entering a
minimal basic fee and informing case worker
they were not claiming it. Fee calculation
has prevented this.