Validate retrial start dates being on or after trial end dates
As part of retrial fee calculation it was found that
alot of `retrial_stared_at` dates were being entered with
a date earlier than the `trial_concluded_at`. This leads
to a negative retrial interval, which in turn, causes the
fee calculator to return a basic fee amount 'Advocate fee`
equivalent to the trial (as if the user was not claiming
a reduction on the retrial).

- the "negative retrial interval" could be a result of
  user error, or attempting to claim no reduction.
- in any event the data should be accurate and it should
  not be possible for retrial to start before a trial ends

This validation has been put in place to enforce the
business/legislative rules AND to enable accurate
calculation of retrial reductions based on retrial
intervals