Fix flaky spec.
Autoloading seems to be initializing the CCR
constant twice during the supplementary feature test.

The result is a fee calc response error in which
the response has no responseJSON. This fixes the
JS that errors as a result, and something that
probably should be catered for in any event. It
therefore also fixes the falkiness of the feature
test, but without idenitfying the autoloading issue
(which does occur on development, at least).

TODO: track down cause of double autoloading of CCR constant
in app:
```
activesupport-5.1.6/lib/active_support/dependencies.rb:457: warning: already initialized constant CCR
activesupport-5.1.6/lib/active_support/dependencies.rb:457: warning: previous definition of CCR was here
```