Modify bill scenario service in line with supplementary claims
Supplementary claims do not require a case type
as the fee/price values are not effected by them.

However, the fee calc api always requires a scenario
(derived from case type) so we must default the
scenario to something. I have chosen to map it
to trial by default (as the most common) claim
case type.