Prevent unique PK error
Seems to be happening when you
create a factory generated misc fee type
after having seeded all fee types.
Because seeds are specifying ids to use
the sequence is not being updated so a
autoincremented sequence id is clashing
with existing id.

Needs changing as part of test seeding perhaps
to update the relevant sequence in line with
the latest/max id +1.