Add advocate final claim API endpoint
The introduction of additional advocate endpoints
means they are better presented as endpoints
in their own namespace.

This creates an advocate "final" claim endpoint
following the namespacing for the other advocate claim
types (interim and supplmenetary).

The intention is to deprecate and eventually remove the
"old" `v1/external_users/claims` endpoint.