Amend vcr configuration
Config was in one place for specs with addendums
for cucumber. Simpler and more consistent to
store in allvcr config in one place `spec/vcr_helper`
and require this for cucumber via `env.rb`.

In addition ensure we are using the custom matcher,
which was not being applied for cucumber tests as a
result for faulty config.