Amending feature tests for unit price calc to account for debounce delay
Since $.debounce(wait_in_milliseconds) means ajax
request is not fired immediately the speed of feature
test steps means ajax may not even have fired before
wait_for_ajax is called, and therefore expectations
may not be met.

An alternative could be to use `patiently`, but this
is not dealing with the actual issue, rather its effects.