Add miscellaneous SAF uplift to supplmentary claims
The Standard appearance fee and its uplift are
both claimable on supplementary claims.

The "Standard appearance fee uplift" (MISAU) is now
and uplift for both the BASAF and MISAF. As a result
the defendant upliftable logic needs changing to handle
this possible two parent scenario.

It is also an uplift for FXSAF, but this is handle via
the fixed fee adapter.