Fix bugs for FactoryBot upgrate 4.8.2 -> 5.0.2
Errors were occuring in this upgrade due to FactoryBot deprecating
the use of statically defined variables in fatories in v5.

**To fix:**

Had to Upgrade to 4.11.1 first. Which gave deprecation warning re
static declarations.

It suggested using `rubocop-rspec` which has the cop:

`FactoryBot/AttributeDefinedStatically`

Which triggers when it sees statically defined attributes in factories.

I included this as a dependency (in Gemfile and .rubocop.yml) , then had to
unignore `spec/` from `.rubocop.yml` and run `rubocop --auto-correct` to
correct static declrations.

Then upgraded to `5.0.2` and rspecs seem to run okay.