Add Spec DI spec for defendant uplifts with two possible parents
With the introduction of the miscellaneous Standard appearance fee
(SAF) there are now three possible parents for the miscellaneous
Standard appearance fee uplift (MISAU) - BASAF, FXSAF and now MISAF.

The BASAF and MISAF need to be differentiated from in order
to avoid conflicts between DI and Fee calculation usage.

The uplift for the fixed fee equivalent (FXSAF) needs considering.