Turn on forgery protection in feature tests (only)
Feature tests appear to occasionally be failing because
logged in user is losing sessions when forgery protection
is off. With it turned on we at least get and
InvalidAuthenticityToken error, which may be an underlying
cause.

This only started occurring when ruby 2.6 and
capybara/site_prims/webdrivers/chromedriver were
bumped??!