Amend Zendesk Stub inline with User-Agent change
Move from 1.17.0 to 1.19.0 changes the User-Agent
header and this broke the existing stub.

I have amended the spec to make this more obvious
when failing. Previously the stubbing mechanism
resulted in an actual request being made and returning
401 (unauthorized) because it was stubbing a request
with specific headers (which then changed).

Now the spec stubs all request to the Zendesk URI
and the spec tests that a specific request was made.