Add mocked invalid state transition error in `transtitoner_spec`
I am not convinced the "actual" tests for
claims with errors being archived successfully
are doing what we would hope them too.

This explcitly raises the kind of error
we are seeing on sentry and tests the logging
and the fact we are swalling the error - to
prevent the batch transister job from failing
entirely.