Add required secrets to archive stale cronjob
live-1 cronjob failing on placeholder
prod environment with live data.

The error received was
```
undefined methods `match` for nil:NilClass
```

This is can be raised by paperclip. Could
also be caused by MI data problems.

Better to provide all env vars that are available
in any environment to be confident the job can work
in any environment. Note that zendesk, google api
and other keys are not available to all environments
and so are excluded.