Return http status 200 if only sidekiq queue has issues
Returning a 502 for Sidekiq queues failures
(DeadSets and Retries) is causing live-1 cluster
to intercept at the nginx level and redirect
to the default/generic bad-gateway page.

In addition the healthcheck is used to tell if the
app is healthy. A failed queued job does not
denote the app as unhealthy in any event.