Get jasmine specs passing
Jasmine was failing to load assets
due to rails version specific logic
in the gem. Rails 6 was not included
for asset pipeline asset compilation.

However `jasmine-gem` master branch
includes rails 6 sprockets/asset-pipeline
compilation and does, therefore work.

This is not an ideal solution but hopefully
version 3.6/4.0 will provide this as an option.