[WIP] add hardship claim casetype-to-stage mapper
Ask user what stage the case reached and map
this this to a case type on the back end. This
is to ensure users select the correct case type
given the complexities of current hardship
claim logic, so they claim the correct amount - as mapped
by the business stakeholders/caseworkers.

TODO:
  - extract case type to own partial
  - add case type to case stage mapper