Prevent seeding of sample users on application load
This has been happening for a long time and should
not have been.

It was caused by autoload of `lib` dir in combination
with the existence `lib/demo_data/demo_seeds.rb` file
which loads the seeds as soon as it is loaded (i.e.
on application boot, due to autoload of lib dir).

To explicitly load sample users use:
```
rake claims:sample_users
```