Further updates (#3304)
* Remove commented out cleaner
Having reviewed the other clenaers, I cannot see the need
for one on this claim type due to its simplicity, happy
to discuss and be corrected!

* load the hardship_fee_type_id specifically
No need to infer from case type * this was a quick fix and will need reviewing next as we need to change the case type anyway! 

* Remove undeeded sidebar entries for LGFS hardship
* Update case stage handling
* Update view for fee fields
* Update case_stage validations for litigator hardship
* Update the feature to use the new case stage name
* Update shared_examples "common litigator claim attributes"
Allow user to skip the requires_trial_dates? check
as this is not needed only for the hardship claim type
Also updated a couple of the descriptions as they did
not match the tests that ere performed internally