Add hardship claim and fee adapters
Hardship claims and their grad fees need to be
injected into Crown court remuneration (CCR)
as AGFS_ADVANCE AGFS_HARDSHIP bill types.

This bill type is not calculated by CCR and, instead
accepts a single amount, excluding VAT.

Current thinking is that this equates to
an aggregation (sum) of all the basic fees
of a hardship claim in CCCD, except Standard appearances
(BASAF), Plea and trial preparation hearings (BAPCM)
and Conferences and views (BACAV). The latter fees
are injected as misc fees or, in the case of Conferences
and views, not injected. Expenses are injected separately
similarly.