Extend timeout for requests
There have been quite a few cloner failures
in the last month (may 2020), 32 of which
display and error message in the logs of
"execution expired".

It is believed this is due to the number
of documents and/or their sizes associated with
the claim being cloned.

Increasing the timeout is a test to see the impact.

A more robust solution would be to background the cloning
process or the file copy aspect of it.