Add panel component helpers using a lib and railtie
This pattern should make moving to a rails extension
gem easier, and make it easy to add components
in a modular fashion.

The GovukComponent::Railtie class mixes in the
helpers at boot time. see [this good blog on the subject](https://blog.capsens.eu/rails-6-boot-sequence-d289b44d2e94)

To get around the fact that railties need to be
intialized early in the boot sequence the railtie
is `require`'d in the `config/application.rb` (just after
normal Gem `require`ing). Once moved to a gem this can
be removed and would simply work as part of the normal
`Bundler.require(*Rails.groups)` call in that file.