Extend allow_nil option to max, min and equal methods
To allow these methods to not fail with nil values
for source if developer decides this is required.

Needed to prevent validation errors or NameError
(NoMethodError) at innapppropriate times in the
lifecycle ofr claim creation.

Example: to allow a CLAR fee to be added before
a defendant with rep order is added. But once
rep order is added we need to validate existence of
CLAR fee types only for claims with representation order
date AFTER (minimum rule) CLAR release date.