Remove Unused material (over 3 hours) quantity validation
We have become aware of a logic issue with the current
implementation of the CLAR “Unused material (over 3 hours)”
fee type for AGFS. The regulations stipulate that the first 3
hours are claimable at the fixed rate in addition to any hours
over 3 at an hourly rate. This means we will be allowing any
quantity to be claimed for this fee type where each unit/hour
should be in excess of the first 3 hours.

example:
  3 and half hours of work for “Unused material” would be claimable
  as two fees: `quantity: 1` of type “Unused material (upto 3 hours)”
  and `quantity: 0.5` of type “Unused material (over 3 hours)”.