Unused materials warning text (#3516)
Add warning text to unused materials over 3 hours

### Why
Mitigate user confusion

From a provider's perspective, if we're splitting 'up to 3 hours/over 3 hours' there's a real possibility for confusion. If a provider needs to claim for 5 hours of unused prep, they're going to be thinking "I enter 5 hours". Given the fee rules though, the system needs the provider to enter 2 hours (the extra hours).

### How
to handle across 3 different views of misc fees
with either select list, radio buttons or check box
options (for the same fee), we are using js to
hide a show the warning on the fee type. This
requires a data attribute to indicate the unique
code which can then be used to display or not.

### TODO
jasmine specs to unit test the JS have not 
been doable in the short time frame.

The eligible misc fee collection requires a
custom data attribute of unique code to determine
various options to display per fee type.

Co-authored-by: Peter Hunt <naseberry@users.noreply.github.com>